% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeConsensus}
\alias{computeConsensus}
\title{Compute consensus sequence for a strain.}
\usage{
computeConsensus(strain, ref, anno, thr = 0.01)
}
\arguments{
\item{strain}{- chr str, name of the wb strain contained in externalId.}

\item{ref}{- DNAStringSet from dbFasta().}

\item{anno}{- data frame from dbAnno().}
}
\description{
\code{computeConsensus} returns consensus sequence for a strain.
}
