% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainGPHMM.R
\name{computeCounts}
\alias{computeCounts}
\title{Compute counts for emissions and transitions.}
\usage{
computeCounts(gphmm)
}
\arguments{
\item{gphmm}{- output of computegphmm with arg output='long'.}
}
\description{
\code{computeCounts} returns a list with emission and transition counts.
}
\examples{
gphmmOut <- computegphmm('ATCG', 'ATG', output = 'long')
computeCounts(gphmmOut)
}
