% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gphmm.R
\name{computeDelta}
\alias{computeDelta}
\title{Compute insertion and deletion rates, i.e., probability to transition from
match/mismatch state to insertion and deletion state.}
\usage{
computeDelta(mat, qv)
}
\arguments{
\item{mat}{- matrix with counts for insertions and deletions for each read.}

\item{qv}{- vector with quality values for the reads.}
}
\description{
\code{computeDelta} returns list of coefficients from logit regression for insertion and deletion rates.
}
