% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainGPHMM.R
\name{computeEps}
\alias{computeEps}
\title{Compute normalized transition probabilities.}
\usage{
computeEps(transition, state = "X")
}
\arguments{
\item{transition}{- named vector with at least elements 'DD','DM','II','IM'.}

\item{state}{- str, "X" or "Y".}
}
\description{
\code{computeEps} returns normalized transition probabilities.
}
