% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gphmm.R
\name{computeQ}
\alias{computeQ}
\title{Compute normalized emission probabilities.}
\usage{
computeQ(emission)
}
\arguments{
\item{emission}{- vector of length at least 4.}
}
\description{
\code{computeQ} returns normalized emission probabilities.
}
