% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findFormat}
\alias{findFormat}
\title{Find format of a file.}
\usage{
findFormat(file)
}
\arguments{
\item{file}{- chr str, name of file to extract the format from.}
}
\description{
\code{findFormat} returns format of file.
}
