% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findLastJson}
\alias{findLastJson}
\title{Find last json files created.}
\usage{
findLastJson(sfx = "gphmm", path = paste0(system.file(package = "gphmm"),
  "/training/"))
}
\arguments{
\item{sfx}{- chr str, suffix of the json file, for example, 'parambt'.}

\item{path}{- chr str, path of the directory to look for.}
}
\description{
\code{findLastJson} returns name of last json created.
}
