% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ShortReadQToShortRead}
\alias{ShortReadQToShortRead}
\title{Transform queries in format ShortReadQ to format ShortRead.}
\usage{
ShortReadQToShortRead(queries)
}
\arguments{
\item{queries}{- ShortReadQ object.}
}
\description{
\code{ShortReadQToShortRead} returns queries in ShortRead format.
}
