% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ShortReadToShortReadQ}
\alias{ShortReadToShortReadQ}
\title{Transform queries in format ShortRead to format ShortReadQ.}
\usage{
ShortReadToShortReadQ(queries)
}
\arguments{
\item{queries}{- ShortRead object.}
}
\description{
\code{ShortReadToShortReadQ} returns queries in ShortReadQ format.
}
