% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitTrainTest}
\alias{splitTrainTest}
\title{Randomly generate sequences.}
\usage{
splitTrainTest(y, proptrain = 0.8, seed = NULL)
}
\arguments{
\item{y}{- any R object for which method length() has been defined.}

\item{proptrain}{- float, percentage to be included in training set.}

\item{seed}{- integer.}
}
\description{
\code{splitTrainTest} returns a DNAStringSet with generated sequences.
}
