% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toFasta}
\alias{toFasta}
\title{Read fasta or fastq file and returns ShortRead format.}
\usage{
toFasta(fPath)
}
\arguments{
\item{fPath}{- chr str, path to file with queries, can be a fasta or fastq file.}
}
\description{
\code{toFasta} returns queries in ShortRead format.
}
