% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toFastq}
\alias{toFastq}
\title{Read fasta or fastq file and returns ShortReadQ format.}
\usage{
toFastq(fPath)
}
\arguments{
\item{fPath}{- chr str, path to file with queries, can be a fasta or fastq file.}
}
\description{
\code{toFastq} returns queries in ShortReadQ format.
}
