% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{traceBack}
\alias{traceBack}
\title{Generate trace back.}
\usage{
traceBack(q, i, j, Tm, Tx, Ty)
}
\arguments{
\item{q}{double, index of the state for the last two nucleotides in the read and reference sequence.}

\item{i}{double, length of the read.}

\item{j}{double, length of the reference sequence.}

\item{Tm}{matrix, GPHMM probabilities at each location in the read and reference sequence in the M state.}

\item{Tx}{matrix, GPHMM probabilities at each location in the read and reference sequence in the insertion state.}

\item{Ty}{matrix, GPHMM probabilities at each location in the read and reference sequence in the deletion state.}
}
\description{
This function returns the  trace back, i.e. the sequence of most probable 
states in the alignment between read x and reference sequence y.
}
