% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gphmm.R
\name{computegphmm}
\alias{computegphmm}
\title{Compute GPHMM log probability.}
\usage{
computegphmm(read, ref, parameters, qv = 20, output = "short")
}
\arguments{
\item{read}{- chr str.}

\item{ref}{- chr str.}

\item{parameters}{- list of GPHMM parameters, can be created by initializeGphmm().}

\item{qv}{- float, quality value.}

\item{output}{- if 'long', output is a list with path, read, ref and log GPHMM proba, else output is just the log GPHMM proba.}
}
\description{
\code{computegphmm} returns GPHMM log probability for a read and a reference sequence.
}
