% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gphmm.R
\name{computeGphmmParam}
\alias{computeGphmmParam}
\title{Compute gphmm parameters from counts.}
\usage{
computeGphmmParam(emiTrans)
}
\arguments{
\item{emiTrans}{- list with emission and transition counts.}
}
\description{
\code{computeGphmmParam} returns a list with gphmm parameters.
}
