% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateRandomSequences}
\alias{generateRandomSequences}
\title{Split randomly any R object for which method length() has been defined into a train set and a test set.}
\usage{
generateRandomSequences(n = 2, meanLen = 10, sdLen = 0, seed = NULL,
  ncores = NULL)
}
\arguments{
\item{n}{- int, number of sequences to generate.}

\item{meanLen}{- float, mean of the length distribution (gaussian) of the generated sequences.}

\item{sdLen}{- float, sd of the length distribution (gaussian) of the generated sequences.}

\item{seed}{- int, when the same seed and parameters are used, exact same sequences are generated.}

\item{ncores}{- int, number of cores to use.}
}
\description{
\code{generateRandomSequences} returns a list with indices for train and test sets.
}
